--- ongoing-papers.ts
@@ -1,43 +1,0 @@
-import { type IUpcomingPaper } from "@/interface";
-
-const papers: IUpcomingPaper[] = [
-  {
-    _id: "6708fd8002a75017a4f08759",
-    subject: "Discrete Mathematics and Graph Theory [BMAT205L]",
-    slots: ["A2", "B2"],
-  },
-  {
-    _id: "670980523ec3fdad83b2d211",
-    subject: "Compiler Design [BCSE307L]",
-    slots: ["C1", "D1"],
-  },
-  {
-    _id: "670a105e6272bcf9da4e2362",
-
-    final_url:
-      "https://res.cloudinary.com/dtorpaj1c/image/upload/v1728712773/papers/ifkrnjgwdudtev9rxpki.pdf",
-    thumbnail_url:
-      "https://res.cloudinary.com/dtorpaj1c/image/upload/w_400,h_400,c_fill/v1728712773/papers/ifkrnjgwdudtev9rxpki.jpg",
-    subject: "Complex Variables and Linear Algebra [BMAT201L]",
-    slots: ["A1", "B1"],
-    year: "2023",
-    exam: "CAT-2",
-    semester: "Fall Semester",
-    campus: "Vellore",
-  },
-  {
-    _id: "67097e7b3ec3fdad83b2d205",
-    final_url:
-      "https://res.cloudinary.com/dtorpaj1c/image/upload/v1728675439/papers/pyyqotz1mzuh2wmq1k9s.pdf",
-    thumbnail_url:
-      "https://res.cloudinary.com/dtorpaj1c/image/upload/w_400,h_400,c_fill/v1728675439/papers/pyyqotz1mzuh2wmq1k9s.jpg",
-    subject: "Computer Networks [BCSE308L]",
-    slots: ["C2", "D2"],
-    year: "2023",
-    exam: "CAT-2",
-    semester: "Fall Semester",
-    campus: "Vellore",
-  },
-];
-
-export default papers;
