+++ src/components/PapersCarousel.tsx
@@ -14,6 +14,7 @@ import {
 import Autoplay from "embla-carousel-autoplay";
 import { chunkArray } from "@/util/utils";
 import { Skeleton } from "@/components/ui/skeleton";
+import SkeletonPaperCard from "@/components/SkeletonPaperCard";
 
 function PapersCarousel() {
   const [displayPapers, setDisplayPapers] = useState<IUpcomingPaper[]>([]);
@@ -78,23 +79,7 @@ function PapersCarousel() {
                 chunkSize === 4 ? "grid-cols-2 grid-rows-2" : "grid-cols-4"
               } gap-4 lg:auto-rows-fr`}
             >
-              {Array.from({ length: chunkSize }).map((_, idx) => (
-                <div
-                  key={idx}
-                  className="cursor-pointer rounded-sm border-2 border-[#734DFF] bg-[#FFFFFF] text-black shadow-lg transition duration-150 ease-in-out hover:bg-[#EFEAFF] dark:border-[#36266D] dark:bg-[#171720] dark:text-white hover:dark:bg-[#262635]"
-                >
-                  <div className="border-b-2 border-[#453D60] p-2">
-                    <Skeleton className="h-6 w-24 rounded-md" />
-                  </div>
-                  <div className="flex flex-col justify-between p-4">
-                    <Skeleton className="mb-4 h-6 w-32 rounded-md" />
-                    <div className="flex gap-2">
-                      <Skeleton className="h-7 w-16 rounded-full" />
-                      <Skeleton className="h-7 w-16 rounded-full" />
-                    </div>
-                  </div>
-                </div>
-              ))}
+              <SkeletonPaperCard length={chunkSize} />
             </CarouselItem>
           ) : (
             chunkedPapers.map((paperGroup, index) => {
