+++ src/components/PinnedPapersCarousel.tsx
@@ -17,6 +17,7 @@ import AddPapers from "./AddPapers";
 import Autoplay from "embla-carousel-autoplay";
 import { chunkArray } from "@/util/utils";
 import { StoredSubjects } from "@/interface";
+import SkeletonPaperCard from "./SkeletonPaperCard";
 
 function PinnedPapersCarousel({
   carouselType = "upcoming",
@@ -54,14 +55,10 @@ function PinnedPapersCarousel({
         localStorage.getItem("userSubjects") ?? "[]",
       ) as StoredSubjects;
 
-      console.log("Fetching papers for subjects:", storedSubjects);
-
       const response = await axios.post<{ subject: string; slots: string[] }[]>(
         "/api/user-papers",
         storedSubjects,
       );
-
-      console.log("Fetched papers:", response.data);
 
       const fetchedPapers = response.data;
 
@@ -131,23 +128,7 @@ function PinnedPapersCarousel({
                   chunkSize === 4 ? "grid-cols-2 grid-rows-2" : "grid-cols-4"
                 } gap-4 lg:auto-rows-fr`}
               >
-                {Array.from({ length: chunkSize }).map((_, idx) => (
-                  <div
-                    key={idx}
-                    className="cursor-pointer rounded-sm border-2 border-[#734DFF] bg-[#FFFFFF] text-black shadow-lg transition duration-150 ease-in-out hover:bg-[#EFEAFF] dark:border-[#36266D] dark:bg-[#171720] dark:text-white hover:dark:bg-[#262635]"
-                  >
-                    <div className="border-b-2 border-[#453D60] p-2">
-                      <Skeleton className="h-6 w-24 rounded-md" />
-                    </div>
-                    <div className="flex flex-col justify-between p-4">
-                      <Skeleton className="mb-4 h-6 w-32 rounded-md" />
-                      <div className="flex gap-2">
-                        <Skeleton className="h-7 w-16 rounded-full" />
-                        <Skeleton className="h-7 w-16 rounded-full" />
-                      </div>
-                    </div>
-                  </div>
-                ))}
+                <SkeletonPaperCard length={chunkSize} />
               </CarouselItem>
             ) : (
               chunkedPapers.map((paperGroup, index) => {
